refactor user agreement consent notification theme and widget
