+++ collaboration-suite-core-ui/src/androidTest/java/com/kaleyra/collaboration_suite_core_ui/UserDataConsentAgreementNotificationTest.kt
@@ -2,7 +2,7 @@ package com.kaleyra.collaboration_suite_core_ui
 
 import org.junit.Test
 
-class EULANotificationTest {
+class UserDataConsentAgreementNotificationTest {
 
     @Test
     fun test() {
