+++ collaboration-suite-core-ui/src/main/res/values/themes.xml
@@ -81,4 +81,33 @@
 
     <style name="KaleyraCollaborationSuiteUI.Theme.DayNight" parent="KaleyraCollaborationSuiteUI.Theme.Day" />
 
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Base" parent="Theme.MaterialComponents.DayNight.NoActionBar">
+        <item name="colorPrimaryVariant">@android:color/transparent</item>
+        <item name="colorPrimaryDark">?attr/colorPrimaryVariant</item>
+        <item name="fontFamily">@font/kaleyra_font</item>
+        <item name="android:fontFamily">@font/kaleyra_font</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Day" parent="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Base">
+        <item name="colorPrimary">@color/kaleyra_terms_colorPrimary</item>
+        <item name="colorOnPrimary">@color/kaleyra_terms_colorOnPrimary</item>
+        <item name="colorSecondary">@color/kaleyra_terms_colorSecondary</item>
+        <item name="colorOnSecondary">@color/kaleyra_terms_colorOnSecondary</item>
+        <item name="android:colorBackground">@color/kaleyra_terms_colorBackground</item>
+        <item name="colorOnBackground">@color/kaleyra_terms_colorOnBackground</item>
+        <item name="android:windowLightStatusBar" tools:targetApi="m">true</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Night" parent="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Base">
+        <item name="colorPrimary">@color/kaleyra_terms_colorPrimary_night</item>
+        <item name="colorOnPrimary">@color/kaleyra_terms_colorOnPrimary_night</item>
+        <item name="colorSecondary">@color/kaleyra_terms_colorSecondary_night</item>
+        <item name="colorOnSecondary">@color/kaleyra_terms_colorOnSecondary_night</item>
+        <item name="android:colorBackground">@color/kaleyra_terms_colorBackground_night</item>
+        <item name="colorOnBackground">@color/kaleyra_terms_colorOnBackground_night</item>
+        <item name="android:windowLightStatusBar" tools:targetApi="m">false</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.DayNight" parent="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Day"/>
+
 </resources>
