+++ collaboration-suite-phone-ui/src/main/java/com/kaleyra/collaboration_suite_phone_ui/userdataconsentagreement/UserDataConsentAgreementWidget.kt
@@ -1,19 +1,19 @@
-package com.kaleyra.collaboration_suite_core_ui.eula
+package com.kaleyra.collaboration_suite_phone_ui.userdataconsentagreement
 
 import android.content.Context
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.kaleyra.collaboration_suite_core_ui.R
-import com.kaleyra.collaboration_suite_core_ui.databinding.KaleyraEulaBinding
+import com.kaleyra.collaboration_suite_phone_ui.databinding.KaleyraUserDataConsentAgreementWidgetLayoutBinding
 
-class EulaWidget @JvmOverloads constructor(
+class UserDataConsentAgreementWidget @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = R.attr.kaleyra_rootLayoutStyle
 ) : ConstraintLayout(context, attrs, defStyleAttr) {
 
-    private val binding: KaleyraEulaBinding by lazy { KaleyraEulaBinding.inflate(LayoutInflater.from(context), this, true) }
+    private val binding: KaleyraUserDataConsentAgreementWidgetLayoutBinding by lazy { KaleyraUserDataConsentAgreementWidgetLayoutBinding.inflate(LayoutInflater.from(context), this, true) }
 
     fun setTitle(text: String) {
         binding.kaleyraTitle.text = text
