+++ collaboration-suite-phone-ui/src/main/res/layout/kaleyra_user_data_consent_agreement_widget_layout.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    android:theme="@style/KaleyraCollaborationSuiteUI.TermAndConditionsTheme.DayNight">
+    android:theme="@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.DayNight.Phone">
 
     <com.google.android.material.textview.MaterialTextView
         android:id="@+id/kaleyra_title"
