--- collaboration-suite-phone-ui/src/main/res/values/terms_and_conditions.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources xmlns:tools="http://schemas.android.com/tools">
-    <declare-styleable name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme">
-        <attr name="kaleyra_rootLayoutStyle" />
-        <attr name="kaleyra_titleStyle" />
-        <attr name="kaleyra_subtitleStyle" />
-        <attr name="kaleyra_positiveButtonStyle" />
-        <attr name="kaleyra_negativeButtonStyle" />
-    </declare-styleable>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar">
-        <item name="colorPrimaryVariant">@android:color/transparent</item>
-        <item name="colorPrimaryDark">?attr/colorPrimaryVariant</item>
-        <item name="fontFamily">@font/kaleyra_font</item>
-        <item name="android:fontFamily">@font/kaleyra_font</item>
-        <item name="kaleyra_rootLayoutStyle">@style/KaleyraCollaborationSuiteUI.TermAndConditionsTheme.RootLayout</item>
-        <item name="android:scrollViewStyle">@style/KaleyraCollaborationSuiteUI.TermAndConditionsTheme.ScrollView</item>
-        <item name="kaleyra_titleStyle">@style/KaleyraCollaborationSuiteUI.TextView.Title.TermsAndConditions</item>
-        <item name="kaleyra_subtitleStyle">@style/KaleyraCollaborationSuiteUI.TextView.Subtitle.TermsAndConditions</item>
-        <item name="kaleyra_positiveButtonStyle">@style/KaleyraCollaborationSuiteUI.TermAndConditionsTheme.FilledButton</item>
-        <item name="kaleyra_negativeButtonStyle">@style/KaleyraCollaborationSuiteUI.TermAndConditionsTheme.TextButton</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.Day" parent="KaleyraCollaborationSuiteUI.TermAndConditionsTheme">
-        <item name="colorPrimary">@color/kaleyra_terms_colorPrimary</item>
-        <item name="colorOnPrimary">@color/kaleyra_terms_colorOnPrimary</item>
-        <item name="android:colorBackground">@color/kaleyra_terms_colorBackground</item>
-        <item name="colorOnBackground">@color/kaleyra_terms_colorOnBackground</item>
-        <item name="android:windowLightStatusBar" tools:targetApi="m">true</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.Night" parent="KaleyraCollaborationSuiteUI.TermAndConditionsTheme">
-        <item name="colorPrimary">@color/kaleyra_terms_colorPrimary_night</item>
-        <item name="colorOnPrimary">@color/kaleyra_terms_colorOnPrimary_night</item>
-        <item name="android:colorBackground">@color/kaleyra_terms_colorBackground_night</item>
-        <item name="colorOnBackground">@color/kaleyra_terms_colorOnBackground_night</item>
-        <item name="android:windowLightStatusBar" tools:targetApi="m">false</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.DayNight" parent="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.Day"/>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.RootLayout" parent="">
-        <item name="android:padding">@dimen/kaleyra_dimen_space40</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.ScrollView" parent="">
-        <item name="android:fadeScrollbars">false</item>
-        <item name="android:scrollbars">vertical</item>
-        <item name="android:layout_marginBottom">@dimen/kaleyra_dimen_space48</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TextView.Title.TermsAndConditions" parent="KaleyraCollaborationSuiteUI.TextView.Title">
-        <item name="android:textColor">?attr/colorOnBackground</item>
-        <item name="android:layout_marginBottom">@dimen/kaleyra_dimen_space24</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TextView.Subtitle.TermsAndConditions" parent="KaleyraCollaborationSuiteUI.TextView.Subtitle">
-        <item name="android:textColor">?attr/colorOnBackground</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.FilledButton" parent="Widget.MaterialComponents.Button">
-        <item name="android:textAppearance">@style/KaleyraCollaborationSuiteUI.TermAndConditionsTheme.ButtonTextAppearance</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.TextButton" parent="Widget.MaterialComponents.Button.TextButton">
-        <item name="android:alpha">0.6</item>
-        <item name="android:textAppearance">@style/KaleyraCollaborationSuiteUI.TermAndConditionsTheme.ButtonTextAppearance</item>
-    </style>
-
-    <style name="KaleyraCollaborationSuiteUI.TermAndConditionsTheme.ButtonTextAppearance" parent="">
-        <item name="android:textSize">@dimen/kaleyra_dimen_text_normal</item>
-        <item name="android:textAllCaps">false</item>
-    </style>
-
-</resources>
