+++ collaboration-suite-phone-ui/src/main/res/values/user_data_consent_agreement.xml
@@ -1,0 +1,72 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Phone">
+        <attr name="kaleyra_rootLayoutStyle" />
+        <attr name="kaleyra_titleStyle" />
+        <attr name="kaleyra_subtitleStyle" />
+        <attr name="kaleyra_positiveButtonStyle" />
+        <attr name="kaleyra_negativeButtonStyle" />
+    </declare-styleable>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.DayNight.Phone" parent="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.DayNight">
+        <item name="kaleyra_rootLayoutStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.RootLayout</item>
+        <item name="android:scrollViewStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.ScrollView</item>
+        <item name="kaleyra_titleStyle">@style/KaleyraCollaborationSuiteUI.TextView.Title.UserDataConsentAgreement</item>
+        <item name="kaleyra_subtitleStyle">@style/KaleyraCollaborationSuiteUI.TextView.Subtitle.UserDataConsentAgreement</item>
+        <item name="kaleyra_positiveButtonStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.FilledButton</item>
+        <item name="kaleyra_negativeButtonStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.TextButton</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Day.Phone" parent="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Day">
+        <item name="kaleyra_rootLayoutStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.RootLayout</item>
+        <item name="android:scrollViewStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.ScrollView</item>
+        <item name="kaleyra_titleStyle">@style/KaleyraCollaborationSuiteUI.TextView.Title.UserDataConsentAgreement</item>
+        <item name="kaleyra_subtitleStyle">@style/KaleyraCollaborationSuiteUI.TextView.Subtitle.UserDataConsentAgreement</item>
+        <item name="kaleyra_positiveButtonStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.FilledButton</item>
+        <item name="kaleyra_negativeButtonStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.TextButton</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Night.Phone" parent="KaleyraCollaborationSuiteUI.UserDataConsentAgreementTheme.Night">
+        <item name="kaleyra_rootLayoutStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.RootLayout</item>
+        <item name="android:scrollViewStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.ScrollView</item>
+        <item name="kaleyra_titleStyle">@style/KaleyraCollaborationSuiteUI.TextView.Title.UserDataConsentAgreement</item>
+        <item name="kaleyra_subtitleStyle">@style/KaleyraCollaborationSuiteUI.TextView.Subtitle.UserDataConsentAgreement</item>
+        <item name="kaleyra_positiveButtonStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.FilledButton</item>
+        <item name="kaleyra_negativeButtonStyle">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.TextButton</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreement.RootLayout" parent="">
+        <item name="android:padding">@dimen/kaleyra_dimen_space40</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreement.ScrollView" parent="">
+        <item name="android:fadeScrollbars">false</item>
+        <item name="android:scrollbars">vertical</item>
+        <item name="android:layout_marginBottom">@dimen/kaleyra_dimen_space48</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.TextView.Title.UserDataConsentAgreement" parent="KaleyraCollaborationSuiteUI.TextView.Title">
+        <item name="android:textColor">?attr/colorOnBackground</item>
+        <item name="android:layout_marginBottom">@dimen/kaleyra_dimen_space24</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.TextView.Subtitle.UserDataConsentAgreement" parent="KaleyraCollaborationSuiteUI.TextView.Subtitle">
+        <item name="android:textColor">?attr/colorOnBackground</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreement.FilledButton" parent="Widget.MaterialComponents.Button">
+        <item name="android:textAppearance">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.ButtonTextAppearance</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreement.TextButton" parent="Widget.MaterialComponents.Button.TextButton">
+        <item name="android:alpha">0.6</item>
+        <item name="android:textAppearance">@style/KaleyraCollaborationSuiteUI.UserDataConsentAgreement.ButtonTextAppearance</item>
+        <item name="android:background">@android:color/transparent</item>
+    </style>
+
+    <style name="KaleyraCollaborationSuiteUI.UserDataConsentAgreement.ButtonTextAppearance" parent="">
+        <item name="android:textSize">@dimen/kaleyra_dimen_text_normal</item>
+        <item name="android:textAllCaps">false</item>
+    </style>
+
+</resources>
